New model algorithms and APIs for scope analysis (#1478)
Previously we identified implicit base models of datatypes and other model children, but we only did this from the
perspective of the child (the "extension"), not from the perspective of the base.  This worked fine until Matter 1.3
when the specification began relying more heavily on the "extended cluster" construct.

This commit adds the notion of a "scope" to the model.  Instead of querying a model for applicable members, you now
query the scope.  The scope is aware of extensions, so this allows you to retrieve the correct set of members in
contexts (such as datatype fields defined in the base cluster) where there is otherwise no explicit reference to the
extension.

CHANGELOG.md contains additional details.

Fixes #1455