Extract protocol init from node init
Fixes a mistake in the implementation of node initialization.

Before this commit, ClusterBehaviors were backed by a backing that managed the cluster server.  This intermingled protocol setup with node initialization.

Now there is still a ServerBacking but creation of ClusterServers is handled independently when the node is brought online.

Primary motivation for fixing this now is that it allows for behaviors to independently contribute components used by the network runtime, which we will use in future commits.