Synchronize node lifecycle transitions
This places protections around node lifecycle transitions with many async steps.  Previously it was possible to
trigger these simultaneously in ways that cause crashes due to interference between them.  They may still be
invoked in parallel but will execute sequentially.  Protected methods are available for composing the operations
when the mutex is already held.