Event persistence (#1502)
* Extract protocol init from node init

Fixes a mistake in the implementation of node initialization.

Before this commit, ClusterBehaviors were backed by a backing that managed the cluster server.  This intermingled protocol setup with node initialization.

Now there is still a ServerBacking but creation of ClusterServers is handled independently when the node is brought online.

Primary motivation for fixing this now is that it allows for behaviors to independently contribute components used by the network runtime, which we will use in future commits.

* Event persistence

Adds the term "occurrence" to differentiate between "classes of events that I can subscribe to" and "historical instances of said event".

Splits the functionality of EventHandler into storage (EventStore) and logic (OccurrenceManager).  Adds volatile and non-volatile EventStore implementations.  Extends the algorithm for pruning old events a bit.

By default OccurrenceManager operates in volatile mode.

Primarily needed now for testing, but created with public API as reliable event delivery seems like table stakes for nodes playing a role in security.

* Update CHANGELOG.md

---------

Co-authored-by: Ingo Fischer <github@fischer-ka.de>