Improve construction error handling (#1532)
When component initialization fails due to an error in a dependency, we previously did not report the error in the
dependency because this could lead to duplicate stack traces.

The downside is that if the process crashes before the root cause is reported then it goes unreported.  This commonly
happens with unhandled exceptions.  And even if reported, the error message would potentially be in a separate location
in the log that you'd need to hunt for.

This commit makes it so we attach the original cause to "dependency initialization" errors.  To avoid redundant traces
it includes logic to track which errors have had their trace logged.  For tertiary reporting of the error we then omit
the stack trace so the only duplication is of the message.

Also includes a few additional tests and a fix for late addition of behaviors with async initialization logic.

Co-authored-by: Ingo Fischer <github@fischer-ka.de>