Logging Optimization and Restructuring (#1585)
* Adds a "Keylike" Log renderer

"Keylike" means that the key is suppressed and the value is rendered like a key.

* Add flags format function

This gets an object of values and the keys of all "truthy" values will be added as space delimited list. "keylikeFlags" directly renders it with a keylike presentation.

* Enhance dict() to suppress undefined values

* Add missing import

* Add logContext to Exchange and fill it

This commits adds a "logContext" to the Exchange#send code flow and fills it with values. Ideally Messagetypes and such are determined automatically from the protocolId and opCode

* Streamline logging

This commit streamlines logging in the two Messengers and simplifies it.

* Allows to disable MRP logic

Allows to disable MRP logic wehn sending exchange data - this is used when pushing out subscription data on shutdown where we do not want to have a full MRP retransmission cycle, so we just push the message without waiting for ack and just wait for Success for a shorter time to not block shutdown. In this case oushing submission infos is "by chance" anyway, so that's ok

* Optimize Session logging

* Optimize Mdns Logging

* changelog

* Adjust tests

* remove duplicate changelog entry

* Address review feedback

* Address review feedback