Full scope search for named references (#1600)
Adds support for default values that are named references.  For any property with a default value that is a named reference, a managed struct will return the referenced value until set explicitly.

Adds logic that searches for properties by name in any owner of a managed value.  This allows conformance expression constraints and default values to reference values by name so long as they are visible in any object in the ownership hierarchy.  Previously we only supported referencing values that were siblings.

I believe this is the last feature we need to support 100% of the data model semantics used in the Matter specifications (as of Matter 1.4).