BLE commissioning optimizations (#1605)
* Always re-discover BLE devices

... with this we can try to correct weird noble states at least for now

* Adjust state when rediscovering.

When we rediscovered the device then we know it is not connected and can reset the stat because noble does not do this for now. Additionally add a needed parameter which was not supported by BleScanner till now.

* Adjust scanning logic to mimic noble behavior

Sometimes "scanStart is emitted" without we asked for because of internal reasons and scanStop is missed, so ignore these events

* Adjust scanning logic to mimic noble behavior

Sometimes "scanStart is emitted" without we asked for because of internal reasons and scanStop is missed, so ignore these events

* We want duplicated discoveries to reset status

* Remember instances to not recreate them

* Added basic support for Concurrent and Non-Concurrent commissioning flows

* Refactor connection handling

... mainly to adress all special cases that can happen with noble where communication hangs or connections do not work and such. tested manually.

Logging stays a bit verbose for now to see more when issues from users come in

* Changelog

* Add eslint exceptions as before