Stores MDNS host information for faster reuse with new SRV announcements
Host records were not persisted when they were used already or such, so we needed to re-ask them. in commissioning processes the hostname likely stays the same when the device switches from commissioning to operative mode, so we could reuse the host information easiely.

The PR also adds expiry logic for active query data and details and for these stored data.

it also adds a bit more logging to be likely removed before we publish 0.12 but allows better validation of the fix with the affected users.