Trigger event listeners before attribute listeners for Subscriptions (#1658)
* Trigger event listeners before attribute listeners for Subscriptions

In fact happened events usually lead to a new "final attribute status" which is also reported. To mimic this a bit better we should trigger event listeners before attribute listeners. This could halp a lot with e.g. Switch events where events report presses, but the switch position is reported as "0".

* fix tests

* Update CHANGELOG.md

---------

Co-authored-by: lauckhart <greg@lauckhart.com>