Minor log fixes (#1677)
* Format IPv6 URLs with port correctly

IPv6 addresses must be wrapped in brackets before the port is appended.
Even though the printed URLs are only informative, it makes it easier to
see where the IP ends and the port begins.

* Avoid blank error description when message is empty

A RetransmissionLimitReachedError error with no message was just
yielding a blank line after the colon. Fall back to stringifying the
error in its entirety in this case.

---------

Co-authored-by: Ingo Fischer <github@fischer-ka.de>