Allow to establish and work with a PASE session only
Allows to establish "just" a PASE session to a device and then use this channel to get an InteractionClient to work with.
This is there just for special cases and custom commissioning cases - and also used in testing.
IN fact this Pase channel is not official registered anywhere because in fact there is no peer ... I would leave it that way for now if we see more use cases