Allow working with non commissioned nodes in system
For testing I needed to also communicate with devices not officially commissioned or close such connections. These changes allow that nd allow to bypass a bit.
That's mainly used in testing and we should discuss such use cases with ServerNode how to really do that because it opens some questions :-) For now I think that way it is ok because mainly used for testing - thats why also not officially documented too much and not in Changelog.