WebSocket Controller testing and Controller Enhancements (#1780)
* Splitup TestDevice

... because COntroller (at least now) is different regarding features and requirements. I assume we will bring it together again later but separated for now

* Define a CommandHandler Interface

* Implement CommandHandler for Legacy Controller

This class implements all needed logic and really interacts with the CommissioningController. WHen later moving to ServerNode we just need to re-implement this part and map data accordingly.

* ImplementChiptool WebSocket Handler

This class is the heart of the controller test. It opens the WebSocket and receives chip-tool compatible messages, does all the conversion magic and such and interfaces with ControllerHandler. Then it also re-converts the response to be chip compatible

* Add Custom Commissioning Flow

... does not set Fabric Label. This is used by Legacy Controller Handler

* Add Controller WebSocket Test App

* Add Controller WebSocket Sniffer as Tooling support

* Add ws dependency

* Cleanup

* Typing optimization

* Forward Facility name in leg call

... this is needed to include this also in the chip response, so ignored in most other places

* Allow PairedNode with non reconnectable client

In speci cases we need to allow an Interacron client that is not reconnectable. because this is special case it logs a warning. Also not added to changelog

* Allow custom NodeId, CA and CommissioningFlow impl

Introduces the option to specify a custom Root-Node-ID, an own CertificateAuthority or a CommissioningFlow implementation.

This was requested in Discord and is also needed for the Controller testing. Several properties of ControllerCommissioningFlow were changed from private to protected to allow inherited classes to use it

* Allow to establish and work with a PASE session only

Allows to establish "just" a PASE session to a device and then use this channel to get an InteractionClient to work with.
This is there just for special cases and custom commissioning cases - and also used in testing.
IN fact this Pase channel is not official registered anywhere because in fact there is no peer ... I would leave it that way for now if we see more use cases

* Only consider not-closed channels as available

Make sure to not consider channels in a pre-closing state as ok and existing.

* Treat generated and received StatusResponseErrors differently

We had a mix here and there were StatusResponseErrors that we generated were also send to the device which was wrong

* Parses and Return attribute/event Status from DataReports

Also parses and return AttributeStatus and EventStatus from DataReports and expose via InteractionCLient.

This was missing and now also needed by testing.
I decided to not "throw on error" because there could be multiple and  also valid responses

* Enhance invoke handling

Invoke method in InteractionClient got two enhancements:
* a parameter was added to skip data validation. This will send the message with the provided data for cases where maybe schema is wrong or testing wants to test error cases.
* Add automatic handling for FabricScoped commands to set the fabricIndex. Because as per spec it is ignored anyway we set the "no fabric" identifier.

* Make attributes, events and commands available via Ids

Enhance ClusterClient to also allow to access commands, events or attributes by their ID. This is not typed and so mainly relevant for cases where also cluster is requested by id. The ServerNode implementation will also allow this. Thats just convenience and used in testing

* Inform about subscription updates after the data where received

... this makes more sense

* Hide irrelevant Dgram error

When Dgram is not listening but closed we were logging ann error. prevent this.

* Add Todo

* Allow working with non commissioned nodes in system

For testing I needed to also communicate with devices not officially commissioned or close such connections. These changes allow that nd allow to bypass a bit.
That's mainly used in testing and we should discuss such use cases with ServerNode how to really do that because it opens some questions :-) For now I think that way it is ok because mainly used for testing - thats why also not officially documented too much and not in Changelog.

* Changelog

* Add missing line

* Fix tests

* [execute-chiptests-long] [rebuild-chip] Add test execution via matter.js controller

* [execute-chiptests-long] [rebuild-chip] Fix tests

* [execute-chiptests-long] [rebuild-chip] Fix tests

* [execute-chiptests-long] [rebuild-chip] Try run in docker build container

* [execute-chiptests-long] [rebuild-chip] Path test file

* [execute-chiptests-long] [rebuild-chip] try to patch network

* [execute-chiptests-long] [rebuild-chip] another try - set sysctl setting

* [execute-chiptests-long] [rebuild-chip] another try

* [execute-chiptests-long] [rebuild-chip] another try

* [execute-chiptests-long] [rebuild-chip] another try

* [execute-chiptests-long] [rebuild-chip] another try

* [execute-chiptests-long] [rebuild-chip] split and rebuild normal executables

* [execute-chiptests-long] [rebuild-chip] Address review feedback

* [execute-chiptests-long] [rebuild-chip] different build way again

* [execute-chiptests-long] [rebuild-chip] try these build options

* [execute-chiptests-long] [rebuild-chip] last fixes

---------

Co-authored-by: lauckhart <greg@lauckhart.com>