ColorControl moved to Transitions ... including enhancements (#1859)
* LeveControl: Streamline and Remove default features

We should also here not enable any features by default. And Also name as the others - there was a miss in the other PR

* Centralize rate field checks

* LevelControl: Allow remainingTime to be not present

This field is only allowed when having the Lighting feature enabled, so allow it to be absent.

I hope the !options and !defaults is right way?

* Fix AddWithOverflow logic

* Remove unneeded promise handling

* Add support for cyclic attributes to Transition

Cyclic attributes are the onles like Hue that wrap over at the defined min/max values. These are transitioned using the pre-calculated distance instead of min/max

* Adjust RemainingTime reports rules

Remembers the last published value and considers spec rules based on this. Additionally, prevents "too short actions " (<1s) from being reported at all as requested by spec

* Typos and use data directly

Do not use eventsOf and stateOf because not available

* We need to wait for the first interval trigger to lower the data and time ... else we are transitioning too fast. Thats ok for reality because only after a short time stuff changes and commands should be implemented otherwise anyway

* Only set new value if it has changed

* Adjust CC to Transitions

* Add CC cyclic transition test

* Adjust LVL test

* rename tests npm script

* adjusts tests

* changelog

* linter

* disable CC 9.1/2/3 in non-docker-run

* adjust tests a bit

* small rounding fix

* Update Transitions.ts

Typos & formatting for Transitions

* address review feedback

* oki, disable tests with loop till clarification

* optimize tests

---------

Co-authored-by: lauckhart <greg@lauckhart.com>