Address edge case on reconnect
WHen the session establishment and also subscription was successfull and the device then did not had all data to be collected it could happen that we start communication that needs a longer time to fail - and this potentially multiple times leading to a long delay. If this was longer than the reconnect time then the init was still in progress and so reconnect failed. Edge case.