First Refactor-Step: Interactionserver merge and read wire-in (#1970)
* Merge InteractionServer

This commit merges TransactionalInteractionServer and the InteractionServer together, adjusts tests as needed for this (because now node is main usage) and adjusts all relevant imports.

* Fix ACL checks for unknown endpoints

Even for unknown endpoints we should check ACLs first because access could be not allowed at all or otherwise limited. So we build a fallback ACL context

* Fix Wildcard Bit definition for AttributeTypeProtocol

The reserved Bit 3 was and oversight and so all bits after that were shifted by one leading to wrong behavior

* Enhance ReadResult definitions for Attributes

We also need info about cluster version, tlv-schema and if the dat are fabricSensitive. So we enhhance that structure for now until we also move encoding into new code paths

* Adjust AttributeResponse

* expose more data for encoding
* check if state value is undefined to only report really used attributes
* Add location for ACl checks
* Build a fallback location when cluster is unknown
* send the attribute value instead all data of state

* Read can also be a sync iterator

* Replace attribute read logic by ServerInteraction.read

... and convert the data how the current encoding logic needs it

* Adjust test

* Changelog

* linter

* fix global Attribute wildcard check

* Disable mandatory field validation in encoder for Behavior data

* Make reader async again

* Add DataModelPath.none