Subscription Wiring to Attribute/EventResponse (#2000)
* Added stateChange event to Datasource

The event is fired with the state property names that have changed and is sent before other $Changed events

* Fix wrong call

... never used so far, but error handler should not be "then"

* Add added event to OccurrenceManager

* Refactor AttributeResponse to be extensible.

One main change is to have a "process" method to actually do the logic instead of the constructor. So I also adjusted EventResponse to work the same.

Adjustments needed in order to make AttributeResponse being extensible and create FilteredAttributeResponse to filter process just on a part of data

* Extend ProtocolService

... adding and bubbling up state change events to have one event on "Node" level and we need maps name -> id

* Move report convert logic

For now better move the conversion logic into a place where it is more near to the parts that require it.

* Move ServerSubscription to node package

... and update for new logic using new readers.

* Update tests

* Remove eventsToName Map

* Rename Filterd -> AttributeSubscriptionResponse

* Respect quieter events and some optimizations

This is intermediate and a follow up commit will enhance more

* Move attribute name to ID map to RootSupervisor

* Add dispose handling for cluster state

* Fix name typo

* Prevent sending unwanted empty data reports

e.g. while data are there but filtered out for fabric scoping or such and so ending in empty report.

This solution is intermediate and should be fixed later correctly. Left TODO

---------

Co-authored-by: lauckhart <greg@lauckhart.com>