Write Refactoring and Subscription Fix (#2013)
* Remove unused legacy code

* Rename classes

* Export type for AttributeData

* introduce Interaction Session type

basically is AccessControl.Session + transaction property

* Split read/write access in ClusterProtocol

* rename a count field

* Simplify Read DSL types

I think it is enough to have an Attribute/Event status struct with an optional cluster status. Also because it was basically not really wired for now ... I also saw use cases for invoke here so far ... so we could also remove it till we find a case :-)

* Add inspect support for listIndex field

* verify some Transaction things

* ReadState also needs a session

* Add AttributeWriteResponse handler

* Wire in AttributeWriteResponse handler

* No CancellablePromise anymore

* Add Resource and no Participant

* Define Write DSL

... not working because of value encoding topic

* Adjust tests

* Adjust and fix subscriptions

Subscriptions are adjusted to the latest Read and context changes.
Additionally we are now making sure to watch for changes also during activation timeframe.

* See if first basic tests are working

* Finalize

* Add todo for multi message writes

* Seems change means touched and ot changed

Spec is not really clearly defining what a "change" is, but test expects also "unchanged" entries being reported as "changed". see also https://github.com/CHIP-Specifications/connectedhomeip-spec/issues/11671

* Add some safeguards because owning fabric cound be "0"

* Re-enable ACL checks because we are now in sync with SDK

But we also patch ACE 2.2 for now becauseof https://github.com/project-chip/connectedhomeip/issues/33735

* Add comments for behavior

But we also patch ACE 2.2 for now because of https://github.com/project-chip/connectedhomeip/issues/33735

* make sure to Rollback on errors

* ENhancve edit options for test patching

* Adjust tests

* Adjust tests

* Do not send ACL writes chunked on client side

* Patch + reactivate AccessControlCluster test to updated version