Support for without closures
Refactors transaction execution to allow for use of transactions without a closure (a la `act()`).  This is a precursor step to finishing off aspects of the new protocol implementation.

The `act` pattern is nice because it guarantees transactions are cleaned up when stack frames exit.  However, it does have a few downsides:

* It adds a minimum of two additional stack frames (one for the call to `act`, one for the `actor` function)

* It is incompatible with generators.  It would be possible to proxy generators similarly to how we handle promises but this would be complicated and create additional GC overhead

* Async support means the stack guarantee isn't actually all that robust.  A motivated dev can simply stash the transaction/context into a variable in a higher scope and hold `act()` open with a promise

So... This refactor adds an additional pattern to support transaction management with a more traditional requirement of manual release of the transaction.  It involves an `open()` call with a `Transaction.Finalization` interface that makes it convenient to convey success and/or failure to the transaction.

`MaybeAsync` makes this slightly more complicated than e.g. `using` (in the explicit resource management sense).  But it's not too bad...  See the new implementation of `Transaction.act()` for demonstration of the pattern.

In addition to new `Transaction` plumbing, this commit also:

* Converts `OnlineContext` to use the new code pattern.  Serves as a POC and removes a couple of stack frames.  I did not yet convert `OfflineContext`

* Removes `ActionTracer`.  It has largely remained unused.  We can create a replacement for it when we have a use case

* Performs various trickery to remove frames to make stacks less convoluted.  Not directly related to above but similar mindset and I got tired of looking at the super deep stacks in the debugger

* Adds error diagnostic support for `SuppressedError`.  I didn't actually end up using explicit resource management but this is necessary for nominal support of `using` so worthwhile to keep

* Fixes a few (hypothetical) edge case bugs that I encountered while shuffling transaction code

* Disables one of the LevelControlServer tests.  See TODOs in that code for rationale