Make MDNS Scanner aware criteria from using nodes
The nodes in the system can tell the MDNS scanner criteria they are interested in, so that scanning and storing data can be more focussed on them.

Server nodes basically do not need anything, unless they search for actively.
Client nodes are likely interested in all commissionable devices and the operational devices of their own fabric. So limit to that.

If no criteria is active then all MDNS traffic is skipped.