Apply fabric scoped limiting only to managed entries (#2084)
* Apply fabric scoped limiting only to managed entries

When there was a list with primitives in a fabric-scoped struct then the included list also got handled list if it is fabric scoped. This indeed did not worked and so the list was empty.
This change makes sure that only managed entities (which means "no primitives") are

So this change now in general only applies fabric scooped management when also a managed structure.

* Enhance isObject

* revert isObject checks