Introduce FabricGroupsManager
The FabricsGroupManager class contains an operational optimized view on the Group keys and groups of one fabric. It is added to the fabric and also handles the persistence of the operational group message counters. The content of the class is initialized by the relevant data owing cluster, so group key data itself are not persisted.