Add utility method `IndexedSet#mapOf`
Oftentimes we have sets of things that have a primary lookup key.  We have the choice of storing them in a `Map<key, Thing>` or in a `BasicSet<Thing>`.

In simple cases `Map` is fine, but if we need to lookup based on other fields we either need to maintain a second index or we need to do a slow O(n) search.  And niceties such as `BasicSet`'s events are unavailable.

`BasicSet` OTOH supports the same lookup patterns, plus observation, plus arbitrary secondary indices.  But it does not offer the same level of convenience as does a standard JS `Map` for key/value access.

The `mapOf` utility method in this commit removes this dilemma.  It allows you to obtain a `Map` view of a `BasicSet`.

For example, in `FabricManager` we have:

```ts
    readonly #fabrics = new Map<FabricIndex, Fabric>();
```

But then we manually implement things like `added`/`deleted` events.  And have awkward access patterns like this.#fabrics.values() and where, given a `Fabric`, we must use `Fabric#fabricIndex` to find the entry in the map.

Instead, with this new method we can do something like:

```ts
    const FabricManager extends BasicSet<Fabric> {
        #fabricsByIndex = this.mapFor("fabricIndex");
        /* ...other code here... */
    }
```

This gives us the benefit of both `Map` and `BasicSet` interfaces with no further effort or performance penalty.