Add Groups Server (and partial Client) implementation (#2104)
* Split out membership logic

This change allows to set the needed memberships of an interface as needed separately from the creation. This is needed for Groups to dynamically add or remove them when groups are added or removed.

* Add microsecond support to Time

* Add ipv6BytesToString

* Add evented Map

* Split up Session classes

... and adjust "SecureSession" as abstract parent class for both  group and Unicast sessions

* Adjust/Enhance Datatype classes

* When data are already received no need to calculate a timeout

* ActionContext access check session-free and group introduction

This change reworks acl checks of the Action-Context (or better AccessControl.Session) to not need to rely on the session instance anymore, but make sure all operative details come from context. It also introduces the logic needed to transport all details for group sessions into ACL checks and report the relevant ACL Manager parts

* Introduce SecureGroupSession

This commit introduces a class for a SecureGroupSession and add relevant methods in SessionManager to retrieve or create group Session instances. The instances are stored by Source Node Id

* Add Group Interaction Server logic

Basically add Attribute Write and Command invoke support and forbid the rest.

* Introduce FabricGroupsManager

The FabricsGroupManager class contains an operational optimized view on the Group keys and groups of one fabric. It is added to the fabric and also handles the persistence of the operational group message counters. The content of the class is initialized by the relevant data owing cluster, so group key data itself are not persisted.

* Implement Group key management Cluster

* Implement Groups Cluster

* Initialize Group network on start and events

The ServerNetworkRuntime interfaces with the FabricGroupsManager to initialize and adjust Networking as needed.

* Add Groups on Message codec level

... and check several other things more near to spec

* Add Groups on Exchange level

* Basic Client and PeerSet group support

* Test adjustments

* translate comments

* linter

* [execute-controllertests] Linting and reworks

* [execute-controllertests] Storage context handling

* [execute-controllertests] Enable some more group tests

* [execute-controllertests] Set Fabric storage

* [execute-controllertests] Enhance ChipTool simulator for groups

* [execute-controllertests] Fix InteractionClient group check

* [execute-controllertests] Fix Init of sending group session

* [execute-controllertests] Adjust response suppression for tests

* [execute-controllertests] Scenes are still not testable :-)

* [execute-controllertests] Address Review feedback - Session refactoring

* [execute-controllertests] Address Review feedback #2

* [execute-controllertests] Tweaks imports for more specific

* [execute-controllertests] Next review adjustments

* [execute-controllertests] Next review adjustments

* [execute-controllertests] Final fixes

* [execute-controllertests] Final fixes

* [execute-controllertests] Final fixes

* [execute-controllertests] Remove some logging

* [execute-controllertests] last restructure

* [execute-controllertests] fix