Finalize interactionBegin/end and datasourceCHanged events
Streamlined the event definitions and also how they are wired. InteractionEnd is now tied to the exchange "starting to close" which is after all logic is done. That makes more sense than to wait until client received and acked the message (including potential retransmissions if not received.
The datasourceCHanged event is now also writed in 8was missing) and fires after all single change events