Generate classes without eval
Previously we used eval to give the output class a name which is useful for debugging and diagnostics.  We now create
the class manually as a function assigned to an object property.  This conveys the appropriate name to the VM.

When creating an anonymous ES6 class the named property hack doesn't work so we must build the class using old-school
function + prototype.  This makes the code more complicated but has the additional benefit of making things less
confusing for transpilers.

The primary goal here was to support environments where eval is not fully supported.  As a side benefit this may avoid
JIT deoptimizations and other inefficiencies caused by use of eval.