Introduce DCL types and basic fetch functions
This PR is preparing the ground for DCL usage, so it is not wired in.
Types are taken from Spec but adjusted as it is in reality by the DCL servers, because some fields differ. A basic fetch client is also prepared to allow to read data from the DCL servers.

The real wiring-in comes later