Alias QuickCrypto to crypto in ReactNativeCrypto
This fixes typing so @ts-expect-error is no longer required...  This in turn turned up a reference to undefined variable "digest" in `createHkdfKey`.  I believe it should be a SHA 256 HMAC so set to that.

Also attempt to work around typing of the QuickCrypto export that appears incorrect and fix misleading error from StandardCrypto when the input subtle implementation is undefined.