Slim down byte types
Turns out "BufferSource" is too narrow as it precludes shared buffers.  The alternative is the verbose AllowSharedBufferSource.  I think this is a mistake of TS authors as the more common case is AllowSharedBufferSource.  They should have made a DisallowSharedBufferSource instead.

But anyway.  This creates a type "Bytes" that aliases to "AllowSharedBufferSource" and uses that almost everywhere we take bytes.  That allows us to work with Uint8Array instead of Uint8Array<ArrayBuffer>.

It's "almost everywhere" because there are some APIs where shared buffers actually are disallowed (Blob and Crypto).  In those cases we use new `Bytes.exclusive()` which simply asserts that the buffer is not shared and then casts.